:bug: Fix: 오류 수정
